Completed, i hope

+++ app/templates/profile_ads.php
@@ -10,7 +10,7 @@
     <meta name="viewport"
           content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Профиль</title>
+    <title><?=$title?></title>
 
     <link href="../styles/reset.css" rel="stylesheet" type="text/css">
     <link href="../styles/page.css" rel="stylesheet" type="text/css">
@@ -33,14 +33,13 @@
                 <img src="images/no_user_avatar.svg" alt="">
             </div>
 
-            <div class="profile__user__name">Имя Фамилия</div>
+            <div class="profile__user__name"><?= $data['username'] ?></div>
 
         </div>
         <div class="profile__menu__list">
             <ul>
                 <a href="../profile/my_ads"><li>Мои объявления</li></a>
                 <a href="../profile/watched"><li>Просмотренные</li></a>
-                <a href=""><li>Сообщения</li></a>
                 <a href="../profile/favorites"><li>Избранные</li></a>
                 <a href="../profile/archive"><li>Архив</li></a>
                 <a href="../profile/settings"><li class="profile__menu__list__settings">Настройки</li></a>
@@ -67,13 +66,13 @@
                 </h1>
             </div>
 
-            <?php for($i = 0; $i < count($data); $i++){ ?>
+            <?php for($i = 0; $i < count($data)-1; $i++){ ?>
 
             <div class="ads__ad">
                 <a href="../ad/<?=$data[$i]['id']?>"><img class="ads__ad__mainImage" src="<?=$data[$i]['image_url']?>"/></a>
                 <div class="ads__ad__info">
                     <div class="ads__ad__info__price-and-title">
-                        <span>
+                        <span class="main-info__price">
                             <?php echo $data[$i]['min_price'] . "-" . $data[$i]['max_price'];?>
                         </span>
                         <span id="ad__title">
@@ -85,19 +84,29 @@
                     </div>
                 </div>
                 <div class="ads__ad__settings">
-                <?php if($title === "Просмотренные"): ?>
-                    <a class="ads__ad__setting__img" href="../profile/deleteWatched/<?=$data[$i]['id']?>"><img class="ads__ad__setting__img" src="../images/delete_watched.svg" alt=""></a>
+                <?php if($title === "Избранные"): ?>
+                    <a class="ads__ad__setting__img" href="../profile/favorites/delete/<?=$data[$i]['id']?>"><img class="ads__ad__setting__img" src="../images/delete_watched.svg" alt=""></a>
                 <?php else:{?>
                     <img class="ads__ad__setting__img" src="../images/adSetting_icon.svg" alt="">
                     <object class="ad__settings">
                         <ul>
-                            <a href=""><li>Редактировать</li></a>
+                            <?php if($title === "Просмотренные"):?>
+                                <a href="../create_ad/edit/<?=$data[$i]['id']?>"><li>В избранное</li></a>
+                            <?php else:?>
+                                <a href="../create_ad/edit/<?=$data[$i]['id']?>"><li>Редактировать</li></a>
+                            <?php endif; ?>
+
                             <?php if($title === "Мои объявления"):?>
                                 <a href="../profile/archiveAd/<?=$data[$i]['id']?>"><li>Архивировать</li></a>
                             <?php elseif($title === "Архивированные"):?>
                                 <a href="../profile/unarchiveAd/<?=$data[$i]['id']?>"><li>Разархивировать</li></a>
                             <?php endif; ?>
+
+                            <?php if($title === "Мои объявления" || $title === "Архивированные") : ?>
                                 <a href="../profile/deleteAd/<?=$data[$i]['id']?>"><li>Удалить</li></a>
+                            <?php elseif($title === "Просмотренные") : ?>
+                                <a href="../profile/deleteWatched/<?=$data[$i]['id']?>"><li>Удалить</li></a>
+                            <?php endif; ?>
                         </ul>
                     </object>
 
@@ -122,6 +131,5 @@
 
 
 <script src="../scripts/profile_ads.js"></script>
-<script src="../scripts/header.js"></script>
 </body>
 </html>
